pyproject.toml
[tool.poetry]
name = "blossom"
version = "0.0.0"
description = "The site!"
authors = ["Grafeas Group Ltd. <devs@grafeas.org>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
django-widget-tweaks = "^1.4"
stripe = "^2.48.0"
requests = "^2.22.0"
pygments = "^2.4"
djangorestframework = "^3.11.0"
djangorestframework-api-key = "^1.4.1"
psaw = "^0.0.7"
redis = "^3.3.11"
social-auth-app-django = "^3.1.0"
python-dotenv = "^0.10.3"
dpath = "^1.4.2"
timeout-decorator = "^0.4.1"
psycopg2-binary = "^2.8.4"
drf-yasg = "^1.17"
toml = "^0.10.0"
bugsnag = "^3.6"
slackclient = "^2.5"
pyyaml = "^5.3"
praw = "^7.1.0"
django-filter = "^2.3.0"
gunicorn = "^20.1.0"
honeycomb-beeline = "^2.17.0"
typed-ast = "^1.5.0"
django-revproxy = {git = "https://github.com/hcpuk/django-revproxy.git"}
Markdown = "^3.3.6"
django-ipware = "^4.0.0"
django-mathfilters = "^1.0.0"
django-decorator-include = "^3.0"
pytest-django = "^3.7"
pytest-cov = "^2.8"
pytest-mock = "^1.13"
pytest-parallel = "^0.1.1"
shiv = "^1.0.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
better-exceptions = "^0.2.2"
black = "^22.3.0"
flake8 = { version = "^3.7.9", allow-prereleases = true }
flake8-annotations = "^2.0.1"
flake8-black = "^0.1.1"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
flake8-variables-names = "^0.0.3"
pep8-naming = "^0.9.1"
pre-commit = "^2.4.0"
seed-isort-config = "^2.1.1"
isort = "^4.3.21"
coveralls = "^1.11.1"
mimesis = "^4.0.0"
blossom-wrapper = { git = "https://github.com/GrafeasGroup/blossom-wrapper.git", branch = "master" }
poetry2setup = "^1.0.0"

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git/
  | \.mypy_cache
  | \.github
  | \.pytest_cache
  | \.venv
  | build
  | dist
  | htmlcov
  | blossom.egg-info
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "blossom.settings.testing"
python_files = "tests.py test_*.py *_tests.py"
addopts = "--nomigrations --cov=. --cov-report=html --cov-report=term:skip-covered -p no:warnings"

[tool.poetry.plugins."console_scripts"]
"blossom" = "blossom.main:main"
